Delete api working, merge api into main.py
