+++ datasets/HMRC_RTS/main.py
@@ -6,201 +6,123 @@
 #       extension: .py
 #       format_name: light
 #       format_version: '1.5'
-#       jupytext_version: 1.4.2
+#       jupytext_version: 1.7.1
 #   kernelspec:
 #     display_name: Python 3
 #     language: python
 #     name: python3
 # ---
 
-# ## HMRC Regional Trade Statistics
-#
-# Transform to Tidy Data.
-#
-# The source data is available from https://www.uktradeinfo.com/Statistics/RTS/Documents/Forms/AllItems.aspx in a series of zip files.
-#
-# Each zip file contains fixed-width formatted text files following a layout described in https://www.uktradeinfo.com/Statistics/RTS/Documents/RTS%20Detailed%20data%20information%20pack.pdf. Each row is has two measures: net mass in tonnes and statistical value in £1000's. We're assuming each observation has one measure, so split these  out into separate files.
-
-# +
-from gssutils import *
 import json
+import pandas as pd
+from gssutils import *
 
-def left(s, amount):
-    return s[:amount]
+# # Motion due to no current API-based scraper
 
-scraper = Scraper(json.load(open('info.json'))['landingPage'])
-display(scraper.dataset.landingPage)
-scraper
-# -
+import requests
+from cachecontrol import CacheControl
+from cachecontrol.caches.file_cache import FileCache
+from cachecontrol.heuristics import ExpiresAfter
 
-if scraper.dataset.title.endswith('(RTS'):
-    scraper.dataset.title = scraper.dataset.title + ')'
-scraper.dataset.title
+# Object manages the session
+sess = requests.session()
+cached_sess = CacheControl(sess, cache=FileCache('.cache'), heuristic=ExpiresAfter(days=7))
 
-# zipped data files were linked from https://www.uktradeinfo.com/Statistics/RTS/Pages/RTS-Downloads.aspx, but sometimes this appears to close. If so, look for them directly from https://www.uktradeinfo.com/Statistics/RTS/Documents/Forms/AllItems.aspx
 
-import datetime
-now = datetime.datetime.now()
+# Define child of Scraper which rips from API end point and spits out an dataframe with as_pandas()
+class ApiScraper(Scraper):
+    """
+    Inherits the Scraper class and replace the as_pandas() to get the data from the api endpoint, heavy caching requried
+    """
 
-if len(scraper.distributions) == 0:
-    from lxml import html
-    from gssutils.metadata import Distribution
-    from urllib.parse import urljoin
-    from mimetypes import guess_type
-    from urllib.parse import quote
+    def as_pandas(self):
+        url = self.distributions[0].downloadURL
+        page = cached_sess.get(url)
+        contents = page.json()
+        df = pd.DataFrame(contents['value'])
+        # print('Fetching data from {url}.'.format(url=page.url))
+        while '@odata.nextLink' in contents.keys():
+            # print('Fetching more data from {url}.'.format(url=contents['@odata.nextLink']))
+            page = cached_sess.get(contents['@odata.nextLink'])
+            contents = page.json()
+            df = df.append(pd.DataFrame(contents['value']))
+        return df
 
-    page_uri = 'https://www.uktradeinfo.com/Statistics/RTS/Documents/Forms/AllItems.aspx'
-    page = scraper.session.get(page_uri)
-    tree = html.fromstring(page.text)
-    for anchor in tree.xpath("//a[contains(@href, 'zip')]"):
-        dist = Distribution(scraper)
-        dist.title = anchor.text.strip()
-        dist.downloadURL = urljoin(page_uri, quote(anchor.get('href'), safe='/%'))
-        dist.mediaType, encoding = guess_type(dist.downloadURL)
-        scraper.distributions.append(dist)
-    for anchor in tree.xpath("//a[contains(@href, 'Rtsweb " + str(int(left(str(now.year),2)) - 1) + "')]"):
-        dist2 = Distribution(scraper)
-        dist2.title = anchor.text.strip()
-        dist2.downloadURL = urljoin(page_uri, quote(anchor.get('href'), safe='/%'))
-        dist2.mediaType, encoding = guess_type(dist2.downloadURL)
-        scraper.distributions.append(dist2)
-        
 
-for dist in scraper.distributions:
-    print(dist.downloadURL)
+# For the remainder of the augmentation process, another method
+def fetch_api_to_df(url: str) -> pd.DataFrame():
+    page = cached_sess.get(url)
+    contents = page.json()
+    df = pd.DataFrame(contents['value'])
+    # print('Fetching data from {url}.'.format(url=page.url))
+    while '@odata.nextLink' in contents.keys():
+        # print('Fetching more data from {url}.'.format(url=contents['@odata.nextLink']))
+        page = cached_sess.get(contents['@odata.nextLink'])
+        contents = page.json()
+        df = df.append(pd.DataFrame(contents['value']))
+    return df
 
-# +
-from zipfile import ZipFile, is_zipfile
-from io import BytesIO, TextIOWrapper
 
-out = Path('out')
-out.mkdir(exist_ok=True, parents=True)
-extracted_files = []
 
-observations_file = out / 'observations.csv'
-if observations_file.exists():
-    observations_file.unlink()
-header = True
-
-# For each distribution, open the zipfile and put each source in turn into a dataframe
-# for each source we're looking to create one "value" output, and one "mass" output
-# since the open files are in text file they are processed differently, should go back and change it to unzip 
-#the closed files and process them all as txt which would be tidier but for now this will work
-
-for distribution in scraper.distributions:
-    if distribution.downloadURL.endswith('zip'):
-        with ZipFile(BytesIO(scraper.session.get(distribution.downloadURL).content)) as zip:
-            for name in zip.namelist():
-                with zip.open(name, 'r') as quarterFile:
-                    quarterText = TextIOWrapper(quarterFile, encoding='utf-8')
-                    table = pd.read_fwf(quarterText, widths=[6, 1, 2, 1, 3, 2, 1, 2, 9, 9], names=[
-                        'Period',
-                        'Flow',
-                        'HMRC Reporter Region',
-                        'HMRC Partner Geography',
-                        'Codalpha',
-                        'Codseq',
-                        'SITC Section',
-                        'SITC 4',
-                        'Value',
-                        'Netmass'
-                    ], dtype=str)
-
-                    # Generic changes that apply to both "mass" and "value" outputs
-                    table['Period'] = table['Period'].map(lambda x: f'quarter/{x[2:]}-Q{x[0]}')
-                    table['Flow'] = table['Flow'].map(lambda x: 'exports' if x == 'E' else 'imports')
-
-                    table['HMRC Partner Geography'] = table.apply(
-                        lambda x: x['Codseq'] if x['Codseq'][0] != '#' else x['Codalpha'],
-                        axis=1)
-                    assert table['SITC Section'].equals(table['SITC 4'].apply(lambda x: x[0]))
-                    table.drop(columns=['Codalpha', 'Codseq', 'SITC Section'], inplace=True)
-
-                    table.rename(columns={'Flow':'Flow Directions'}, inplace=True)
-
-                    table['Status'] = 'closed'
-
-                    #Flow has been changed to Flow Direction to differentiate from Migration Flow dimension
-
-                    # Output the mass observations for this input file
-                    mass = table.drop(columns=['Value'])
-                    mass['Measure Type'] = 'net-mass'
-                    mass['Unit'] = 'kg-thousands'
-                    mass.rename(columns={'Netmass': 'Value'}, inplace=True, index=str)
-                    mass.to_csv(observations_file, header=header, mode='a', index=False)
-                    # only output header row the first time
-                    header = False
-
-                    # Output the value observations for this input file
-                    value = table.drop(columns=['Netmass'])
-                    value['Measure Type'] = 'gbp-total'
-                    value['Unit'] = 'gbp-thousands'
-                    value.to_csv(observations_file, header=header, mode='a', index=False)
-                                        
-    else:
-        with BytesIO(scraper.session.get(distribution.downloadURL).content) as text:
-            quarterText = TextIOWrapper(text, encoding='utf-8')
-            table = pd.read_fwf(quarterText, widths=[6, 1, 2, 1, 3, 2, 1, 2, 9, 9], names=[
-                'Period',
-                'Flow',
-                'HMRC Reporter Region',
-                'HMRC Partner Geography',
-                'Codalpha',
-                'Codseq',
-                'SITC Section',
-                'SITC 4',
-                'Value',
-                'Netmass'
-                ], dtype=str)
-
-            # Generic changes that apply to both "mass" and "value" outputs
-            table['Period'] = table['Period'].map(lambda x: f'quarter/{x[2:]}-Q{x[0]}')
-            table['Flow'] = table['Flow'].map(lambda x: 'exports' if x == 'E' else 'imports')
-
-            table['HMRC Partner Geography'] = table.apply(
-            lambda x: x['Codseq'] if x['Codseq'][0] != '#' else x['Codalpha'],
-            axis=1)
-            assert table['SITC Section'].equals(table['SITC 4'].apply(lambda x: x[0]))
-            table.drop(columns=['Codalpha', 'Codseq', 'SITC Section'], inplace=True)
-
-            table.rename(columns={'Flow':'Flow Directions'}, inplace=True)
-
-            table['Status'] = 'open'
-
-            #Flow has been changed to Flow Direction to differentiate from Migration Flow dimension
-
-            # Output the mass observations for this input file
-            mass = table.drop(columns=['Value'])
-            mass['Measure Type'] = 'net-mass'
-            mass['Unit'] = 'kg-thousands'
-            mass.rename(columns={'Netmass': 'Value'}, inplace=True, index=str)
-            mass.to_csv(observations_file, header=header, mode='a', index=False)
-            # only output header row the first time
-            header = False
-
-            # Output the value observations for this input file
-            value = table.drop(columns=['Netmass'])
-            value['Measure Type'] = 'gbp-total'
-            value['Unit'] = 'gbp-thousands'
-            value.to_csv(observations_file, header=header, mode='a', index=False)
+# # And we begin
 
 # +
-from gssutils.metadata import THEME
-scraper.dataset.family = 'Trade'
-scraper.dataset.theme = THEME['business-industry-trade-energy']
-import os
+infoFileName = 'info.json'
 
-dataset_path = pathify(os.environ.get('JOB_NAME', 'gss_data/trade/' + Path(os.getcwd()).name))
+info    = json.load(open(infoFileName))
+scraper = ApiScraper(seed=infoFileName)
+cubes   = Cubes(infoFileName)
 
-with open(out / 'observations.csv-metadata.trig', 'wb') as metadata:
-    metadata.write(scraper.generate_trig())
-
-csvw = CSVWMetadata('https://gss-cogs.github.io/family-trade/reference/')
-csvw.create(
-    out / 'observations.csv', out / 'observations.csv-metadata.json', with_transform=True,
-    base_url='http://gss-data.org.uk/data/', base_path=dataset_path,
-    dataset_metadata=scraper.dataset.as_quads(), with_external=False
-)
+scraper.dataset.family = info['families']
 # -
+
+rts = scraper.as_pandas()
+
+# MonthId 
+# Required Pandas > 1.0
+# df['Period'] = pd.to_datetime(df['MonthId'], format='%Y%m').dt.to_period('Q').dt.strftime('/id/quarter/%Y-%q')
+rts['Period'] = pd.PeriodIndex(pd.to_datetime(rts['MonthId'], format='%Y%m'), freq='Q').astype(str)
+rts['Period'] = rts['Period'].apply(lambda x:'/id/quarter/{year}-{quarter}'.format(year=x[:4], quarter=x[-1:]))
+rts['Period'] = rts['Period'].astype('category')
+
+# FlowTypeId
+# https://api.uktradeinfo.com/FlowType (Strip all strings)
+FlowTypes = fetch_api_to_df(url=info['apiEndpoints']['FlowTypeId']).applymap(lambda x: x.strip() if isinstance(x, str) else x)
+for column in FlowTypes.columns[1:]:
+    FlowTypes[column] = FlowTypes[column].astype('category')
+rts = rts.merge(FlowTypes, on='FlowTypeId').drop(labels=['@odata.type'], axis=1)
+del FlowTypes
+
+
+# GovRegionId
+# https://api.uktradeinfo.com/Region
+GovRegions = fetch_api_to_df(url=info['apiEndpoints']['GovRegionId'])
+for column in GovRegions.columns[1:]:
+    GovRegions[column] = GovRegions[column].astype('category').str.strip()
+rts = rts.merge(GovRegions, left_on='GovRegionId', right_on='RegionId').drop(labels=['RegionId'], axis=1)
+del GovRegions
+
+# CountryID
+# https://api.uktradeinfo.com/Country
+Countries = fetch_api_to_df(url=info['apiEndpoints']['CountryId'])
+for column in Countries.columns[1:]:
+    Countries[column] = Countries[column].astype('category')
+rts = rts.merge(Countries, on='CountryId')
+del Countries
+
+
+# CommoditySitc2Id
+# https://api.uktradeinfo.com/SITC
+SITCs = fetch_api_to_df(url=info['apiEndpoints']['CommoditySitc2Id'])
+for column in SITCs.columns[1:]:
+    SITCs[column] = SITCs[column].astype('category')
+rts = rts.merge(SITCs, left_on='CommoditySitc2Id', right_on='CommoditySitcId').drop(labels=['CommoditySitcId'], axis=1)
+
+
+# Add dataframe is in the cube
+cubes.add_cube(scraper, rts, info['title'])
+
+# Write cube
+cubes.output_all()
